Twelfth Commit: Added and Updated Several Exercises and Added Test Cases for ex26, 29, 31, and 33.

+++ src/main/java/ex31/base/HeartRate.java
@@ -54,8 +54,8 @@ public class HeartRate {
     }
   }
 
-  public long getTargetRate(int age, int rate, int i) {
+  public int getTargetRate(int age, int rate, int i) {
     double intensity = i / 100.0;
-    return Math.round((((220 - age) - rate) * intensity) + rate);
+    return (int) Math.round((((220 - age) - rate) * intensity) + rate);
   }
 }
